requirements.txt
azure-mgmt-resource>=1.1.0
azure-mgmt-compute>=2.1.0
azure-mgmt-network>=1.4.0
azure-mgmt-authorization>=0.30.0
